Move types to separate module
